requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.2.3
django-crispy-forms==1.11.2
django-filter==2.4.0
django-simple-history==3.0.0
django-tables2==2.4.0
pycodestyle==2.7.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
